\name{cgalPolygons-package}
\alias{cgalPolygons-package}
\alias{cgalPolygons}
\docType{package}
\title{
\packageTitle{cgalPolygons}
}
\description{
\packageDescription{cgalPolygons}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cgalPolygons}
\packageIndices{cgalPolygons}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cgalPolygons}

Maintainer: \packageMaintainer{cgalPolygons}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
