% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgalPolygonWithHoles.R
\name{cgalPolygonWithHoles}
\alias{cgalPolygonWithHoles}
\title{R6 class to represent a CGAL polygonWithHoles}
\description{
R6 class to represent a CGAL polygonWithHoles.
}
\examples{

## ------------------------------------------------
## Method `cgalPolygonWithHoles$new`
## ------------------------------------------------

library(cgalPolygons)
pwh <- cgalPolygonWithHoles$new(
  squareWithHole[["outerSquare"]], list(squareWithHole[["innerSquare"]])
)
pwh

## ------------------------------------------------
## Method `cgalPolygonWithHoles$convexParts`
## ------------------------------------------------

library(cgalPolygons)
pwh <- cgalPolygonWithHoles$new(
  squareWithHole[["outerSquare"]], list(squareWithHole[["innerSquare"]])
)
cxparts <- pwh$convexParts()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cgalPolygonWithHoles-new}{\code{cgalPolygonWithHoles$new()}}
\item \href{#method-cgalPolygonWithHoles-convexParts}{\code{cgalPolygonWithHoles$convexParts()}}
\item \href{#method-cgalPolygonWithHoles-print}{\code{cgalPolygonWithHoles$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygonWithHoles-new"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygonWithHoles-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{cgalpolygonWithHoles} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygonWithHoles$new(outerVertices, holes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outerVertices}}{a numeric matrix with two columns, the vertices 
of the outer polygon}

\item{\code{holes}}{a list of numeric matrices, each representing the vertices 
of a hole}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{cgalPolygonWithHoles} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(cgalPolygons)
pwh <- cgalPolygonWithHoles$new(
  squareWithHole[["outerSquare"]], list(squareWithHole[["innerSquare"]])
)
pwh
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygonWithHoles-convexParts"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygonWithHoles-convexParts}{}}}
\subsection{Method \code{convexParts()}}{
Decomposition into convex parts. The polygon must be simple 
  and counter-clockwise oriented.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygonWithHoles$convexParts(method = "triangle")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{the method used: \code{"approx"}, \code{"greene"}, 
or \code{"optimal"}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of matrices; each matrix has two columns and represents 
  a convex polygon.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(cgalPolygons)
pwh <- cgalPolygonWithHoles$new(
  squareWithHole[["outerSquare"]], list(squareWithHole[["innerSquare"]])
)
cxparts <- pwh$convexParts()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygonWithHoles-print"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygonWithHoles-print}{}}}
\subsection{Method \code{print()}}{
Print the \code{cgalPolygonWithHoles} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygonWithHoles$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No value, just prints some information about the polygonWithHoles.
}
}
}
