% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_decagram.R
\docType{data}
\name{decagram}
\alias{decagram}
\title{Decagram}
\format{
A matrix with twenty rows and two columns.
}
\usage{
decagram
}
\description{
A matrix representing a regular decagram. The decagram is 
  counter-clockwise oriented.
}
\keyword{datasets}
