% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_monster.R
\docType{data}
\name{monster}
\alias{monster}
\title{Monster head}
\format{
A list with two fields: a matrix \code{"outer"} and a list 
  \code{"holes"}, containing three matrices.
}
\usage{
monster
}
\description{
A list representing the head of a monster as a polygon with
  holes.
}
\keyword{datasets}
