% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pentagram.R
\docType{data}
\name{pentagram}
\alias{pentagram}
\title{Pentagram}
\format{
A matrix with ten rows and two columns.
}
\usage{
pentagram
}
\description{
A matrix representing a regular pentagram.
}
\keyword{datasets}
