% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_squareWithHole.R
\docType{data}
\name{squareWithHole}
\alias{squareWithHole}
\title{Square with hole}
\format{
A list of two matrices with two columns and four rows.
}
\usage{
squareWithHole
}
\description{
A list representing a square with a square hole 
  (\code{"outerSquare"} and \code{"innerSquare"}).
}
\keyword{datasets}
