% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_star.R
\docType{data}
\name{star}
\alias{star}
\title{Star}
\format{
A matrix with sixteen rows and two columns.
}
\usage{
star
}
\description{
A matrix representing an eight-pointed star. It is 
  counter-clockwise oriented.
}
\keyword{datasets}
