% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgalPolygon.R
\name{cgalPolygon}
\alias{cgalPolygon}
\title{R6 class to represent a CGAL polygon}
\description{
R6 class to represent a CGAL polygon.
}
\examples{

## ------------------------------------------------
## Method `cgalPolygon$new`
## ------------------------------------------------

library(cgalPolygons)
ptg <- cgalPolygon$new(pentagram)
ptg

## ------------------------------------------------
## Method `cgalPolygon$area`
## ------------------------------------------------

library(cgalPolygons)
ptg <- cgalPolygon$new(pentagram)
ptg$area() # should be 5 / sqrt(130 + 58*sqrt(5))
5 / sqrt(130 + 58*sqrt(5))

## ------------------------------------------------
## Method `cgalPolygon$boundingBox`
## ------------------------------------------------

library(cgalPolygons)
ptg <- cgalPolygon$new(pentagram)
plot(ptg$boundingBox(), asp = 1)
polygon(pentagram)

## ------------------------------------------------
## Method `cgalPolygon$plot`
## ------------------------------------------------

library(cgalPolygons)
ptg <- cgalPolygon$new(pentagram)
plot(ptg, lwd = 3, col = "red")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cgalPolygon-new}{\code{cgalPolygon$new()}}
\item \href{#method-cgalPolygon-print}{\code{cgalPolygon$print()}}
\item \href{#method-cgalPolygon-area}{\code{cgalPolygon$area()}}
\item \href{#method-cgalPolygon-boundingBox}{\code{cgalPolygon$boundingBox()}}
\item \href{#method-cgalPolygon-plot}{\code{cgalPolygon$plot()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygon-new"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygon-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{cgalpolygon} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygon$new(vertices)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vertices}}{a numeric matrix with two columns}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{cgalPolygon} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(cgalPolygons)
ptg <- cgalPolygon$new(pentagram)
ptg
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygon-print"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygon-print}{}}}
\subsection{Method \code{print()}}{
Print the \code{cgalPolygon} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygon$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No value, just prints some information about the polygon.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygon-area"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygon-area}{}}}
\subsection{Method \code{area()}}{
Signed area of the polygon.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygon$area()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A number, the signed area of the polygon; it is positive if the 
  polygon is counter-clockwise oriented, negative otherwise.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(cgalPolygons)
ptg <- cgalPolygon$new(pentagram)
ptg$area() # should be 5 / sqrt(130 + 58*sqrt(5))
5 / sqrt(130 + 58*sqrt(5))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygon-boundingBox"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygon-boundingBox}{}}}
\subsection{Method \code{boundingBox()}}{
Bounding box of the polygon.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygon$boundingBox()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A 2x2 matrix giving the lower corner of the bounding box in its 
  first row and the upper corner in its second row.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(cgalPolygons)
ptg <- cgalPolygon$new(pentagram)
plot(ptg$boundingBox(), asp = 1)
polygon(pentagram)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygon-plot"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygon-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the polygon.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygon$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments passed to \code{\link[graphics]{polygon}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No returned value, called for side-effect.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(cgalPolygons)
ptg <- cgalPolygon$new(pentagram)
plot(ptg, lwd = 3, col = "red")
}
\if{html}{\out{</div>}}

}

}
}
