% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgalPolygon.R
\name{cgalPolygon}
\alias{cgalPolygon}
\title{R6 class to represent a CGAL polygon}
\description{
R6 class to represent a CGAL polygon.
}
\examples{

## ------------------------------------------------
## Method `cgalPolygon$new`
## ------------------------------------------------

library(cgalPolygon)
mesh <- cgalMesh$new(meshFile)
rglmesh <- mesh$getMesh()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cgalPolygon-new}{\code{cgalPolygon$new()}}
\item \href{#method-cgalPolygon-print}{\code{cgalPolygon$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygon-new"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygon-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{cgalpolygon} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygon$new(vertices)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vertices}}{if \code{mesh} is missing, must be a numeric matrix with 
three columns}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{cgalPolygon} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(cgalPolygon)
mesh <- cgalMesh$new(meshFile)
rglmesh <- mesh$getMesh()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygon-print"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygon-print}{}}}
\subsection{Method \code{print()}}{
Print a \code{cgalPolygon} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygon$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No value, just prints some information about the polygon.
}
}
}
