% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgalPolygonWithHoles.R
\name{cgalPolygonWithHoles}
\alias{cgalPolygonWithHoles}
\title{R6 class to represent a CGAL polygonWithHoles}
\description{
R6 class to represent a CGAL polygonWithHoles.
}
\examples{

## ------------------------------------------------
## Method `cgalPolygonWithHoles$new`
## ------------------------------------------------

library(cgalPolygons)
pwh <- cgalPolygonWithHoles$new(
  squareWithHole[["outerSquare"]], list(squareWithHole[["innerSquare"]])
)
pwh

## ------------------------------------------------
## Method `cgalPolygonWithHoles$boundingBox`
## ------------------------------------------------

library(cgalPolygons)
pwh <- cgalPolygonWithHoles$new(
  squareWithHole[["outerSquare"]], list(squareWithHole[["innerSquare"]])
)
pwh$boundingBox()

## ------------------------------------------------
## Method `cgalPolygonWithHoles$convexParts`
## ------------------------------------------------

library(cgalPolygons)
pwh <- cgalPolygonWithHoles$new(
  squareWithHole[["outerSquare"]], list(squareWithHole[["innerSquare"]])
)
cxparts <- pwh$convexParts()
pwh$plot(list(), density = 10)
invisible(
  lapply(cxparts, function(cxpart) {
    polygon(cxpart, lwd = 2)
  })
)

## ------------------------------------------------
## Method `cgalPolygonWithHoles$plot`
## ------------------------------------------------

library(cgalPolygons)
pwh <- cgalPolygonWithHoles$new(
  squareWithHole[["outerSquare"]], list(squareWithHole[["innerSquare"]])
)
pwh$plot(
  outerpars = list(lwd = 2), density = 10
)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cgalPolygonWithHoles-new}{\code{cgalPolygonWithHoles$new()}}
\item \href{#method-cgalPolygonWithHoles-boundingBox}{\code{cgalPolygonWithHoles$boundingBox()}}
\item \href{#method-cgalPolygonWithHoles-convexParts}{\code{cgalPolygonWithHoles$convexParts()}}
\item \href{#method-cgalPolygonWithHoles-plot}{\code{cgalPolygonWithHoles$plot()}}
\item \href{#method-cgalPolygonWithHoles-print}{\code{cgalPolygonWithHoles$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygonWithHoles-new"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygonWithHoles-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{cgalpolygonWithHoles} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygonWithHoles$new(outerVertices, holes = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outerVertices}}{a numeric matrix with two columns, the vertices 
of the outer polygon}

\item{\code{holes}}{a list of numeric matrices, each representing the vertices 
of a hole; an empty list is allowed}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{cgalPolygonWithHoles} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(cgalPolygons)
pwh <- cgalPolygonWithHoles$new(
  squareWithHole[["outerSquare"]], list(squareWithHole[["innerSquare"]])
)
pwh
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygonWithHoles-boundingBox"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygonWithHoles-boundingBox}{}}}
\subsection{Method \code{boundingBox()}}{
Bounding box of the polygon with holes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygonWithHoles$boundingBox()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A 2x2 matrix giving the lower corner of the bounding box in its 
  first row and the upper corner in its second row.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(cgalPolygons)
pwh <- cgalPolygonWithHoles$new(
  squareWithHole[["outerSquare"]], list(squareWithHole[["innerSquare"]])
)
pwh$boundingBox()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygonWithHoles-convexParts"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygonWithHoles-convexParts}{}}}
\subsection{Method \code{convexParts()}}{
Decomposition into convex parts of the polygon with holes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygonWithHoles$convexParts(method = "triangle")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{the method used: \code{"triangle"} or \code{"vertical"}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of matrices; each matrix has two columns and represents 
  a convex polygon.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(cgalPolygons)
pwh <- cgalPolygonWithHoles$new(
  squareWithHole[["outerSquare"]], list(squareWithHole[["innerSquare"]])
)
cxparts <- pwh$convexParts()
pwh$plot(list(), density = 10)
invisible(
  lapply(cxparts, function(cxpart) {
    polygon(cxpart, lwd = 2)
  })
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygonWithHoles-plot"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygonWithHoles-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the polygon with holes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygonWithHoles$plot(outerpars, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outerpars}}{named list of arguments passed to 
\code{\link[graphics]{polygon}} for the outer polygon}

\item{\code{...}}{arguments passed to \code{\link[graphics]{polygon}} for the 
holes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No returned value, called for side-effect.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(cgalPolygons)
pwh <- cgalPolygonWithHoles$new(
  squareWithHole[["outerSquare"]], list(squareWithHole[["innerSquare"]])
)
pwh$plot(
  outerpars = list(lwd = 2), density = 10
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cgalPolygonWithHoles-print"></a>}}
\if{latex}{\out{\hypertarget{method-cgalPolygonWithHoles-print}{}}}
\subsection{Method \code{print()}}{
Print the \code{cgalPolygonWithHoles} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cgalPolygonWithHoles$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No value, just prints some information about the polygon.
}
}
}
